Move ValidationUtil to runtime
